FUSETOOLS-2445 - refactor access to stagingrepo preferences
