+++ core/plugins/org.fusesource.ide.camel.model.service.core/META-INF/MANIFEST.MF
@@ -16,8 +16,8 @@ Require-Bundle: org.eclipse.core.resources;bundle-version="3.10.0",
  org.eclipse.m2e.maven.runtime;bundle-version="1.6.2",
  org.jboss.tools.foundation.core;bundle-version="1.2.0",
  org.jboss.tools.foundation.ui;bundle-version="1.2.0",
- org.fusesource.ide.foundation.core;bundle-version="10.0.0",
- org.fusesource.ide.preferences;bundle-version="10.0.0",
+ org.fusesource.ide.foundation.core,
+ org.fusesource.ide.preferences,
  org.eclipse.core.filesystem;bundle-version="1.5.0",
  org.eclipse.core.databinding.observable,
  org.eclipse.jdt.ui,
@@ -27,9 +27,8 @@ Require-Bundle: org.eclipse.core.resources;bundle-version="3.10.0",
  com.fasterxml.jackson.core.jackson-annotations;bundle-version="2.5.0",
  com.fasterxml.jackson.core.jackson-core;bundle-version="2.5.0",
  com.fasterxml.jackson.core.jackson-databind;bundle-version="2.5.0",
- org.fusesource.ide.foundation.core;bundle-version="9.2.0",
- org.fusesource.ide.foundation.ui,
- org.fusesource.ide.preferences;bundle-version="9.2.0"
+ org.fusesource.ide.foundation.core,
+ org.fusesource.ide.foundation.ui
 Bundle-Activator: org.fusesource.ide.camel.model.service.core.internal.CamelModelServiceCoreActivator
 Bundle-ActivationPolicy: lazy
 Export-Package: org.fusesource.ide.camel.model.service.core,
