+++ core/plugins/org.fusesource.ide.preferences/META-INF/MANIFEST.MF
@@ -8,7 +8,8 @@ Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Require-Bundle: org.eclipse.core.resources;bundle-version="3.9.0",
  org.eclipse.core.runtime;bundle-version="3.10.0",
  org.eclipse.draw2d;bundle-version="3.9.100",
- org.eclipse.ui;bundle-version="3.106.0"
+ org.eclipse.ui;bundle-version="3.106.0",
+ org.fusesource.ide.foundation.core;bundle-version="10.1.0"
 Bundle-ActivationPolicy: lazy
 Bundle-Activator: org.fusesource.ide.preferences.Activator
 Export-Package: org.fusesource.ide.preferences;
@@ -16,4 +17,5 @@ Export-Package: org.fusesource.ide.preferences;
    org.eclipse.jface.preference,
    org.eclipse.ui.plugin,
    org.osgi.framework,
-   org.eclipse.core.runtime.preferences"
+   org.eclipse.core.runtime.preferences",
+ org.fusesource.ide.preferences.initializer
