+++ core/plugins/org.fusesource.ide.preferences/OSGI-INF/l10n/bundle.properties
@@ -1,3 +1,4 @@
 #Properties file for org.fusesource.ide.preferences
 Bundle-Vendor = JBoss by Red Hat
 Bundle-Name = Fuse Preferences Plugin
+integration.preferences.project.templates=Staging Repositories
