+++ core/plugins/org.fusesource.ide.preferences/src/org/fusesource/ide/preferences/StagingRepositoryListEditor.java
@@ -8,7 +8,7 @@
  * Contributors:
  * Red Hat, Inc. - initial API and implementation
  ******************************************************************************/
-package org.fusesource.ide.projecttemplates.preferences;
+package org.fusesource.ide.preferences;
 
 import java.util.ArrayList;
 import java.util.List;
