+++ core/plugins/org.fusesource.ide.preferences/src/org/fusesource/ide/preferences/l10n/messages.properties
@@ -1,0 +1,12 @@
+stagingRepositoriesPreferencePageDescription=Provides the ability to specify extra repositories to be added to Camel project. This is useful when working with staging repositories for instance.
+enableStagingRepositories_field=Enable Staging Repositories
+NewRepoDialog_message=Specify details for your new Staging Repository: a unique name and URL must be provided.
+NewRepoDialog_nameinvalid=Repository Name must not be empty and may not include spaces, semi-colons, or commas.
+NewRepoDialog_nameNotUnique=That repository name has already been used in the list of staging repositories. Each name must be unique.
+NewRepoDialog_urlinvalid=Repository URL must be a valid URL.
+NewStagingRepositoryDialogTitle=Specify details for new Staging Repository
+RepositoryName_field=Repository Name:
+RepositoryName_tooltip=Unique name for this Staging Repository.
+RepositoryURL_field=Repository URL:
+RepositoryURL_tooltip=Unique url for this Staging Repository. Must be valid URL.
+stagingRepositoriesList_field=Staging Repositories List
