+++ core/tests/org.fusesource.ide.preferences.tests.integration/META-INF/MANIFEST.MF
@@ -1,0 +1,19 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: %Bundle-Name
+Bundle-SymbolicName: org.fusesource.ide.preferences.tests.integration
+Bundle-Version: 10.1.0.qualifier
+Bundle-RequiredExecutionEnvironment: JavaSE-1.8
+Require-Bundle: org.junit,
+ org.assertj.core;bundle-version="2.1.0",
+ org.eclipse.core.runtime;bundle-version="3.11.1",
+ org.eclipse.core.resources;bundle-version="3.10.1",
+ org.eclipse.jface;bundle-version="3.11.1",
+ org.eclipse.ui.workbench;bundle-version="3.107.1",
+ org.eclipse.ui.ide;bundle-version="3.11.0",
+ org.fusesource.ide.preferences;bundle-version="10.1.0",
+ org.fusesource.ide.camel.tests.util;bundle-version="10.0.0",
+ org.jboss.tools.foundation.core;bundle-version="1.4.0",
+ org.jboss.tools.foundation.ui;bundle-version="1.3.100"
+Bundle-Vendor: %Bundle-Vendor
+Bundle-Activator: org.fusesource.ide.preferences.tests.PreferencesIntegrationTestsActivator
