+++ core/tests/org.fusesource.ide.preferences.tests.integration/OSGI-INF/l10n/bundle.properties
@@ -1,0 +1,3 @@
+#Properties file for org.fusesource.ide.preferences.tests.integration
+Bundle-Vendor = Red Hat
+Bundle-Name = Fuse Tooling Preferences Test Integration
