+++ core/tests/org.fusesource.ide.preferences.tests.integration/build.properties
@@ -1,0 +1,9 @@
+source.. = src/main/resources/,\
+           src/main/java/
+output.. = bin/
+bin.includes = META-INF/,\
+               .,\
+               OSGI-INF/l10n/bundle.properties,\
+               OSGI-INF/
+src.includes = *
+src.excludes = src/
