+++ core/tests/org.fusesource.ide.preferences.tests.integration/infinitest.filters
@@ -1,0 +1,1 @@
+.*IT
