+++ core/tests/org.fusesource.ide.preferences.tests.integration/pom.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <groupId>org.fusesource.ide.core</groupId>
+    <artifactId>tests</artifactId>
+    <version>10.1.0-SNAPSHOT</version>
+    <relativePath>..</relativePath>
+  </parent>
+
+  <artifactId>org.fusesource.ide.preferences.tests.integration</artifactId>
+  <packaging>eclipse-test-plugin</packaging>
+  <name>JBoss Fuse Tooling :: Core :: Tests :: Preferences Integration Tests</name>
+</project>
