+++ core/tests/org.fusesource.ide.preferences.tests/META-INF/MANIFEST.MF
@@ -1,0 +1,11 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: %Bundle-Name
+Bundle-SymbolicName: org.fusesource.ide.preferences.tests
+Bundle-Version: 10.1.0.qualifier
+Fragment-Host: org.fusesource.ide.preferences;bundle-version="10.0.0"
+Bundle-RequiredExecutionEnvironment: JavaSE-1.8
+Require-Bundle: org.junit;bundle-version="4.12.0",
+ org.assertj.core;bundle-version="2.1.0",
+ org.slf4j.api;bundle-version="1.7.2",
+ org.jboss.tools.locus.mockito;bundle-version="1.9.5"
