+++ core/tests/org.fusesource.ide.preferences.tests/OSGI-INF/l10n/bundle.properties
@@ -1,0 +1,2 @@
+#Properties file for org.fusesource.ide.preferences.tests
+Bundle-Name = JBoss Fuse Tooling Preferences Tests
