+++ core/tests/org.fusesource.ide.preferences.tests/build.properties
@@ -1,0 +1,8 @@
+source.. = src/
+output.. = bin/
+bin.includes = META-INF/,\
+               .,\
+               OSGI-INF/l10n/bundle.properties,\
+               OSGI-INF/
+src.includes = *
+src.excludes = src/
