+++ editor/plugins/org.fusesource.ide.projecttemplates/META-INF/MANIFEST.MF
@@ -37,6 +37,7 @@ Require-Bundle: org.eclipse.ui;bundle-version="3.106.0",
  org.jboss.tools.maven.core;bundle-version="1.7.0",
  org.jboss.tools.foundation.core;bundle-version="1.2.1",
  org.jboss.tools.foundation.ui;bundle-version="1.2.1",
+ org.fusesource.ide.preferences;bundle-version="10.0.0",
  org.fusesource.ide.foundation.core;bundle-version="10.0.0",
  org.fusesource.ide.foundation.ui;bundle-version="10.0.0",
  org.fusesource.ide.camel.model.service.core;bundle-version="10.0.0",
@@ -57,8 +58,6 @@ Export-Package: org.fusesource.ide.projecttemplates.adopters,
  org.fusesource.ide.projecttemplates.impl.simple,
  org.fusesource.ide.projecttemplates.internal;x-friends:="org.fusesource.ide.projecttemplates.tests.integration",
  org.fusesource.ide.projecttemplates.maven,
- org.fusesource.ide.projecttemplates.preferences,
- org.fusesource.ide.projecttemplates.preferences.initializer,
  org.fusesource.ide.projecttemplates.util;x-friends:="org.fusesource.ide.projecttemplates.tests.integration,org.fusesource.ide.camel.validation.tests.integration,org.fusesource.ide.launcher.tests.integration",
  org.fusesource.ide.projecttemplates.util.camel;x-friends:="org.fusesource.ide.projecttemplates.tests.integration",
  org.fusesource.ide.projecttemplates.wizards;x-friends:="org.fusesource.ide.projecttemplates.tests.integration",
