+++ editor/plugins/org.fusesource.ide.projecttemplates/OSGI-INF/l10n/bundle.properties
@@ -34,6 +34,6 @@ template.medium.eapspring.description = This template provides a sample Camel ro
 template.medium.cxf.code.first.name = CXF code first
 template.medium.cxf.code.first.description = This template provides a sample Camel route definition which is started by a CXF Webservice call.
 extension-point.name = projectTemplates
-integration.preferences.project.templates=Staging Repositories
+
 template.simple.osespringboot.xml.name=SpringBoot on OpenShift
 template.simple.osespringboot.xml.description=This example demonstrates how to configure Camel routes in Spring Boot via a Spring XML configuration file.
