+++ editor/tests/org.fusesource.ide.projecttemplates.tests.integration/META-INF/MANIFEST.MF
@@ -26,6 +26,7 @@ Require-Bundle: org.junit,
  org.eclipse.m2e.maven.runtime,
  org.jboss.tools.foundation.core;bundle-version="1.3.1",
  org.jboss.tools.foundation.ui;bundle-version="1.3.1",
+ org.fusesource.ide.preferences;bundle-version="10.1.0",
  org.fusesource.ide.foundation.core,
  org.fusesource.ide.foundation.ui,
  org.fusesource.ide.camel.model.service.core,
